[#127] fix: 타입 제거 및 detail page 파일 추가
